Add UK comparison plot
This PR adds markdown for a UK comparison plot based on the daily Rt update from deaths/admissions/cases, plus some explanatory text. I don't know where this is best fit in to the website structure / UK page though.

Sharing this as a first draft, please let me know how best to adapt this so that it slots in elsewhere.